bezig met site/views/arrangements. Ondertussen een rare fout in de backend: in view arrangements geeft de publish knop een -view niet gevonden- 500-error. In de edit layout werkt alles wel normaal

+++ admin/controllers/arrangement.php
@@ -9,26 +9,28 @@ class JexBookingControllerArrangement extends JControllerForm
 		$data = JRequest::getVar("jform");
 	
 		$item_id = $data['id'];
-	
-		$db = JFactory::getDbo();
-		$query = $db->getQuery(true);
-		$query->delete('#__jexbooking_xref_attributes')->where('arr_id = '.$item_id);
-		$db->setQuery($query);
-		$db->query();
-		$db->clear();
-	
-	
-		foreach($data['attribute_id'] as $attribute_id){
+		
+		//eerst kijken of het niet een nieuw item is
+		if($item_id){
 			$db = JFactory::getDbo();
-			$db->clear();
 			$query = $db->getQuery(true);
-			$query->insert('#__jexbooking_xref_attributes');
-			$query->columns('arr_id, attribute_id');
-			$query->values("$item_id, $attribute_id");
+			$query->delete('#__jexbooking_xref_attributes')->where('arr_id = '.$item_id);
 			$db->setQuery($query);
 			$db->query();
+			$db->clear();
+		
+		
+			foreach($data['attribute_id'] as $attribute_id){
+				$db = JFactory::getDbo();
+				$db->clear();
+				$query = $db->getQuery(true);
+				$query->insert('#__jexbooking_xref_attributes');
+				$query->columns('arr_id, attribute_id');
+				$query->values("$item_id, $attribute_id");
+				$db->setQuery($query);
+				$db->query();
+			}
 		}
-	
 		parent::save();
 	}
 }
